+++ examples/vite-react-relay/vite.config.ts
@@ -1,0 +1,8 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+import relay from 'vite-plugin-relay';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+	plugins: [react(), relay],
+});