+++ packages/sdk/src/server/util.ts
@@ -1,5 +1,5 @@
-import { EnvironmentVariable, mapInputVariable } from '../configure/variables';
-import { defaultHost, defaultServerPort, fallbackUrl, isCloud, LoggerLevel, WgEnv } from '../configure/options';
+import { EnvironmentVariable, mapInputVariable, resolveVariable } from '../configure/variables';
+import { defaultHost, defaultServerPort, isCloud, ListenOptions, LoggerLevel, WgEnv } from '../configure/options';
 import { ResolvedServerOptions, ServerOptions, MandatoryServerOptions } from './types';
 
 export const customGqlServerMountPath = (name: string): string => {
@@ -11,19 +11,24 @@ const DefaultServerOptions: MandatoryServerOptions = {
 		host: new EnvironmentVariable(WgEnv.ServerHost, defaultHost),
 		port: new EnvironmentVariable(WgEnv.ServerPort, defaultServerPort),
 	},
-	serverUrl: new EnvironmentVariable(WgEnv.ServerUrl, `http://localhost:${defaultServerPort}`),
+	serverUrl: new EnvironmentVariable(WgEnv.ServerUrl, `http://${defaultHost}:${defaultServerPort}`),
 	logger: {
 		level: new EnvironmentVariable<LoggerLevel>(WgEnv.LogLevel, 'info'),
 	},
+};
+
+export const fallbackServerUrl = (listenOptions: ListenOptions | undefined) => {
+	let port = listenOptions?.port || DefaultServerOptions.listen.port;
+	let host = listenOptions?.host || DefaultServerOptions.listen.host;
+
+	return `http://${resolveVariable(host)}:${resolveVariable(port)}`;
 };
 
 export const serverOptionsWithDefaults = (options?: ServerOptions): MandatoryServerOptions => {
 	return isCloud
 		? DefaultServerOptions
 		: {
-				serverUrl:
-					options?.serverUrl ||
-					new EnvironmentVariable(WgEnv.ServerUrl, fallbackUrl(defaultServerPort, options?.listen)),
+				serverUrl: options?.serverUrl || new EnvironmentVariable(WgEnv.ServerUrl, fallbackServerUrl(options?.listen)),
 				listen: {
 					host: options?.listen?.host || DefaultServerOptions.listen.host,
 					port: options?.listen?.port || DefaultServerOptions.listen.port,